mode addr.go to its own sub-package to reuse
